Add files via upload
The COVID case report is made using POWER BI. As per the report the number of deaths, recovered, and positive cases reported are detailed per each country. Here as per the report sum of Deaths is high in the US and sum of recovered patients is high in Brazil. Positive cases reported from the US seem to be high than other countries.